Bike rental prediction/requirements.txt
streamlit==0.71.0
pandas_profiling
pandas==1.1.3
base58==2.0.1
numpy==1.19.2
pillow==8.0.1
plotly==4.14.1
scikit-learn==0.23.2
lazypredict==0.2.7
seaborn==0.11.1
matplotlib==3.3.3
xgboost==1.1.1
lightgbm==2.3.1
pytest==5.4.3
tqdm==4.56.0